判斷是否為彰化縣網虛擬機，安裝佈景管理模組

+++ xoops.sh
@@ -1,4 +1,4 @@
-#! /bin/bash
+#!/bin/bash
 PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin
 export PATH
 #===============================================================================================
@@ -8,11 +8,14 @@ export PATH
 #   Intro:  https://github.com/xichiou/lamp-xoops
 #===============================================================================================
 
-TADTOOLS_VERSION=3.27
-TADTOOLS_URL="https://campus-xoops.tn.edu.tw/modules/tad_modules/index.php?op=tufdl&files_sn=2010#tadtools_3.27_20190613.zip"
+TADTOOLS_VERSION=3.29
+TADTOOLS_URL="http://campus-xoops.tn.edu.tw/modules/tad_modules/index.php?op=tufdl&files_sn=2042#tadtools_3.29_20190724.zip"
 
-TAD_ADM_VERSION=2.82
-TAD_ADM_URL="https://campus-xoops.tn.edu.tw/modules/tad_modules/index.php?op=tufdl&files_sn=2015#tad_adm_2.82_20190613.zip"
+TAD_ADM_VERSION=2.83
+TAD_ADM_URL="http://campus-xoops.tn.edu.tw/modules/tad_modules/index.php?op=tufdl&files_sn=2049#tad_adm_2.83_20190728.zip"
+
+TAD_THEMES_VERSION=5.6
+TAD_THEMES_URL="https://campus-xoops.tn.edu.tw/modules/tad_modules/index.php?op=tufdl&files_sn=2043#tad_themes_5.6_20190725.zip"
 
 clear
 
@@ -33,27 +36,42 @@ function getIP(){
 
 getIP
 
-rm -rf XoopsCore25-2.5.9
+Public_IP=`curl -s -4 icanhazip.com`
+
+# 檢查是否為虛擬機
+if [ "$Public_IP" == "163.23.200.43" ]; then
+  IP_4=$(echo $IP|cut -d"." -f 4)
+  IP_4=$((IP_4))
+  CHC_PORT=$(($IP_4+20000))
+  URL="${Public_IP}:${CHC_PORT}"
+else
+  CHC_PORT=0
+  URL=$IP
+fi
+
+
+rm -rf XoopsCore25-2.5.10
 rm -rf tadtools
 rm -rf tad_adm
+rm -rf tad_themes
 
-echo "下載 XOOPS 2.5.9 安裝程式並解開..."
-if ! [ -f xoops-2.5.9.zip ];then
-	wget 'http://campus-xoops.tn.edu.tw/modules/tad_uploader/index.php?op=dlfile&cfsn=145&cat_sn=16&name=xoopscore25-2.5.9_tw_20170803.zip' -O xoops-2.5.9.zip
+echo "下載 XOOPS 2.5.10 安裝程式並解開..."
+if ! [ -f xoops-2.5.10.zip ];then
+	wget 'http://campus-xoops.tn.edu.tw/modules/tad_uploader/index.php?op=dlfile&cfsn=1780&cat_sn=16&name=xoopscore25-2.5.10.zip' -O xoops-2.5.10.zip
 	if [ $? -ne 0 ];then
-	  rm -f xoops-2.5.9.zip
+	  rm -f xoops-2.5.10.zip
           echo "主網站下載失敗，改從彰化縣網下載!"
-	  wget 'http://163.23.200.157/site01/XoopsCore25-2.5.9_tw_20170803.zip' -O xoops-2.5.9.zip
+	  wget 'http://163.23.200.9/uploads/xoops-2.5.10.zip' -O xoops-2.5.10.zip
 	  if [ $? -ne 0 ];then
-	    rm -f xoops-2.5.9.zip
+	    rm -f xoops-2.5.10.zip
 	    echo "彰化縣網下載失敗，安裝中斷，請稍後再安裝看看!"
             exit 2;
 	  fi
-        fi
+  fi
 
 fi
-unzip -q xoops-2.5.9.zip
-chown -R apache.apache XoopsCore25-2.5.9
+unzip -q xoops-2.5.10.zip
+chown -R apache.apache XoopsCore25-2.5.10
 
 echo "下載模組並解開： tadtools 工具包..."
 if ! [ -f tadtools_${TADTOOLS_VERSION}.zip ];then
@@ -69,28 +87,19 @@ fi
 unzip -q tad_adm_${TAD_ADM_VERSION}.zip
 chown -R apache.apache tad_adm
 
+echo "下載模組並解開： Tad Themes 佈景管理..."
+if ! [ -f tad_themes_${TAD_THEMES_VERSION}.zip ];then
+	wget $TAD_THEMES_URL -O tad_themes_${TAD_THEMES_VERSION}.zip
+fi
+unzip -q tad_themes_${TAD_THEMES_VERSION}.zip
+chown -R apache.apache tad_themes
+
 #wget --no-check-certificate https://github.com/tad0616/tad_themes/archive/master.zip -O tad_themes.zip
 #unzip -q tad_themes.zip
 #chown -R apache.apache tad_themes-master
 
 
-echo "下載更新並解開： BootStrap4升級補丁"
-if ! [ -f bs4_upgrade.zip ];then
-	wget 'http://120.115.2.90/modules/tad_modules/xoops.php?op=tufdl&files_sn=1845#bs4_upgrade_20190101.zip' -O bs4_upgrade.zip
-fi
-
-if ! [ -d patch ]; then
-	mkdir patch
-fi
-cd patch
-unzip -q -o ../bs4_upgrade.zip
-chown -R apache.apache .
-cp -rf htdocs/* ../XoopsCore25-2.5.9/htdocs/
-echo echo `date "+%Y-%m-%d %H:%M:%S"`>../XoopsCore25-2.5.9/htdocs/uploads/xoops_sn_6.txt
-chown -R apache.apache ../XoopsCore25-2.5.9/htdocs/
-cd ..
-pwd
-cd XoopsCore25-2.5.9
+cd XoopsCore25-2.5.10
 
 # Choose XOOPS site location type
 while true
@@ -101,7 +110,7 @@ do
 	echo -e $MSG_CHOOSE_TYPE_1
 	echo -e $MSG_CHOOSE_TYPE_2
 	echo ""
-	echo "請選擇："
+	echo "請選擇網站類型："
 	echo -e "\t\e[32m1\e[0m. http://${IP}/"
 	echo -e "\t\e[32m2\e[0m. http://${IP}/XOOPS/"
 	read -p "$MSG_INPUT_1" SITE_root_type
@@ -119,6 +128,9 @@ do
 		echo $MSG_INPUT_ONLY "1,2"
 	esac
 done
+
+echo ""
+echo ""
 
 # Choose XOOPS sendmail type
 while true
@@ -146,17 +158,34 @@ if [ $SITE_sendmail_type -eq 1 ]; then
 	sed -i 's/^.*public $SMTP_PORT.*=.*/public $SMTP_PORT =587; \/\/Gmail/g' htdocs/class/mail/phpmailer/class.smtp.php
 fi
 
+echo ""
+echo ""
+
 
 if [ $SITE_root_type -eq 1 ]; then
 	TTIME=`date "+%Y%m%d_%H%M%S"`
+	if [ -f /var/www/html/mainfile.php ]; then
+		echo -e "\n注意!!這台伺服器 /var/www/html 網站根目錄已經有網站，"
+		read -p " 由此腳本幫你搬移原網站到別處，然後再安裝新網站嗎? [y/n]" yn
+		if [ "${yn}" != "Y" ] && [ "${yn}" != "y" ]; then
+			echo 取消安裝
+			exit 3
+		fi
+		echo -e "\n搬移舊網站設定檔到 /var/www/bak_xoops_data_lib/${TTIME}_move"
+		mkdir -p /var/www/bak_xoops_data_lib/${TTIME}_move
+		mv /var/www/xoops_* /var/www/bak_xoops_data_lib/${TTIME}_move
+	fi
+
+	echo -e "\n搬移舊網站到 /var/www/html/${TTIME}_move"
 	mv /var/www/html /var/www/html_${TTIME}_move
 	mv htdocs /var/www/html
-	mv xoops_* /var/www
+	mv -f xoops_* /var/www
 
 
 	cd ..
 	mv tadtools /var/www/html/modules/tadtools
 	mv tad_adm /var/www/html/modules/tad_adm
+	mv tad_themes /var/www/html/modules/tad_themes
 
   if [ -d "/root/DB_Backup/${IP}" ]; then
     echo "Directory /root/DB_Backup/${IP} exists."
@@ -166,7 +195,10 @@ if [ $SITE_root_type -eq 1 ]; then
 	echo ""
 	echo $MSG_SETUP_XOOPS_OK
 	echo ""
-	echo -e $MSG_OPEN_SITE " => http://${IP} " $MSG_TO_FINISH
+	if [ $CHC_PORT -gt 0 ]; then
+		echo -e "\n這台主機是彰化縣網路虛擬機\n"
+  fi
+	echo -e $MSG_OPEN_SITE " => \e[32mhttp://${URL}\e[0m " $MSG_TO_FINISH
 	echo ""
 	echo ""
 fi
@@ -175,7 +207,7 @@ fi
 if [ $SITE_root_type -eq 2 ]; then
 	# Set your XOOPS site location
 	echo $MSG_INPUT_URL
-	echo -e $MSG_YOUR_SITE "http://${IP}/\e[32mXOOPS\e[0m/"
+	echo -e $MSG_YOUR_SITE "http://${URL}/\e[32mXOOPS\e[0m/"
 	read -p "$MSG_CHANGE_DEFAULT" SITE_root
 	if [ -z $SITE_root ]; then
 		SITE_root="XOOPS"
@@ -188,6 +220,8 @@ if [ $SITE_root_type -eq 2 ]; then
 	cd ..
 	mv tadtools /var/www/html/${SITE_root}/modules/tadtools
 	mv tad_adm /var/www/html/${SITE_root}/modules/tad_adm
+	mv tad_themes /var/www/html/${SITE_root}/modules/tad_themes
+
 
 	if [ -d "/root/DB_Backup/${IP}" ]; then
 		echo "Directory /root/DB_Backup/${IP} exists."
@@ -197,8 +231,8 @@ if [ $SITE_root_type -eq 2 ]; then
 	echo ""
 	echo $MSG_SETUP_XOOPS_OK
 	echo ""
-	echo -e $MSG_OPEN_SITE " => http://${IP}/\e[32m${SITE_root}\e[0m/" $MSG_TO_FINISH
-	echo $MSG_STEP_4_14
+	echo -e $MSG_OPEN_SITE " => http://${URL}/\e[32m${SITE_root}\e[0m/" $MSG_TO_FINISH
+	echo -e "請注意：當你開啟上述網址進行安裝到第4步驟時，必須填寫以下資料"
 	echo -e "xoops_data ${MSG_DIR}:\e[32m/var/www/${SITE_root}/xoops_data\e[0m"
 	echo -e "xoops_lib  ${MSG_DIR}:\e[32m/var/www/${SITE_root}/xoops_lib\e[0m"
 	echo ""
